+++ xla/pjrt/c_api_client/pjrt_c_api_client.cc
@@ -874,6 +874,15 @@ absl::StatusOr<std::unique_ptr<PjRtBuffer>> PjRtCApiClient::CreateErrorBuffer(
 
   args.memory = tensorflow::down_cast<PjRtCApiMemorySpace*>(memory)->c_memory();
 
+  absl::flat_hash_map<std::string, xla::PjRtValueType> payload_map;
+  error.ForEachPayload([&](absl::string_view name, const absl::Cord& payload) {
+    payload_map[name] = std::string(payload);
+  });
+  TF_ASSIGN_OR_RETURN(std::vector<PJRT_NamedValue> c_payload,
+                      pjrt::ConvertToPjRtNamedValueList(payload_map));
+  args.payload = c_payload.data();
+  args.num_payload = c_payload.size();
+
   RETURN_STATUS_IF_PJRT_ERROR(c_api_->PJRT_Client_CreateErrorBuffer(&args),
                               c_api_);
 
@@ -2483,6 +2492,34 @@ PjRtCApiExecutable::GetParameterLayouts() const {
     layouts.push_back(std::move(pjrt_layout));
   }
   return layouts;
+}
+
+absl::StatusOr<std::vector<std::vector<absl::string_view>>>
+PjRtCApiExecutable::GetParameterMemoryKinds() const {
+  const PJRT_Api* c_api = pjrt_c_api();
+  if (c_api->pjrt_api_version.major_version == 0 &&
+      c_api->pjrt_api_version.minor_version < 102) {
+    // If the PJRT C API version is too old, fall back to the default
+    // implementation.
+    return Unimplemented(
+        "PJRT_Executable_ParameterMemoryKinds requires PJRT C API version "
+        "0.102 or higher.");
+  }
+  PJRT_Executable_ParameterMemoryKinds_Args args;
+  args.struct_size = PJRT_Executable_ParameterMemoryKinds_Args_STRUCT_SIZE;
+  args.extension_start = nullptr;
+  args.executable = c_executable();
+
+  RETURN_STATUS_IF_PJRT_ERROR(
+      c_api->PJRT_Executable_ParameterMemoryKinds(&args), c_api);
+
+  std::vector<absl::string_view> out;
+  out.reserve(args.num_parameters);
+  for (int i = 0; i < args.num_parameters; ++i) {
+    out.push_back(
+        absl::string_view(args.memory_kinds[i], args.memory_kind_sizes[i]));
+  }
+  return std::vector<std::vector<absl::string_view>>{std::move(out)};
 }
 
 absl::StatusOr<std::vector<std::shared_ptr<const PjRtLayout>>>
@@ -4127,6 +4164,11 @@ absl::StatusOr<std::string> PjRtCApiTopologyDescription::Serialize() const {
 }
 
 absl::StatusOr<uint64_t> PjRtCApiTopologyDescription::Fingerprint() const {
+  if (c_api_->pjrt_api_version.major_version == 0 &&
+      c_api_->pjrt_api_version.minor_version < 101) {
+    TF_ASSIGN_OR_RETURN(std::string serialized, Serialize());
+    return tsl::Fingerprint64(serialized);
+  }
   PJRT_TopologyDescription_Fingerprint_Args args{};
   args.struct_size = PJRT_TopologyDescription_Fingerprint_Args_STRUCT_SIZE;
   args.extension_start = nullptr;